requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
httpx>=0.27.0
openai>=1.0.0
anthropic>=0.40.0
requests>=2.31.0
# content fetching
trafilatura>=1.12.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
yt-dlp>=2026.3.3
curl_cffi>=0.7.0
youtube-transcript-api>=1.2.4
# transcription (requires ffmpeg installed on system)
faster-whisper>=1.0.0
# documents
pdfplumber>=0.11.0
# web UI file uploads
python-multipart>=0.0.9
